+++ backend/config/filesystems.php
@@ -47,6 +47,12 @@ return [
             'report' => false,
         ],
 
+        'rem-uploads' => [
+            'driver' => 'local',
+            'root' => storage_path('app/rem-uploads'),
+            'throw' => false,
+        ],
+
         's3' => [
             'driver' => 's3',
             'key' => env('AWS_ACCESS_KEY_ID'),